+++ doc/src/history.md
@@ -3,8 +3,10 @@
 ## Unreleased
 
 - autoWiring of flake outputs & `mkFlake`
-- home-manager: More unique backup filenames (#97)
 - activate script: add `--dry-run` (#104)
+- home-manager
+  - More unique backup filenames (#97)
+  - Add a default `home.homeDirectory` based on the user's username (#117)
 
 ## 0.2.0 (2024-10-03)